Move modules to the omero_zarr package
