+++ src/omero_zarr/masks.py
