+++ src/omero_zarr/raw_pixels.py
